RIA-3307: Refactored the code to remove the duplicates
