+++ src/test/java/uk/gov/hmcts/reform/iahomeofficeintegrationapi/domain/handlers/presubmit/LeadershipJudgeFtpaDecidedNotificationHandlerTest.java
@@ -1,44 +1,31 @@
 package uk.gov.hmcts.reform.iahomeofficeintegrationapi.domain.handlers.presubmit;
 
-import static java.lang.String.format;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.reset;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static uk.gov.hmcts.reform.iahomeofficeintegrationapi.domain.entities.AsylumCaseDefinition.valueOf;
 import static uk.gov.hmcts.reform.iahomeofficeintegrationapi.domain.entities.ccd.callback.PreSubmitCallbackStage.ABOUT_TO_START;
 import static uk.gov.hmcts.reform.iahomeofficeintegrationapi.domain.entities.ccd.callback.PreSubmitCallbackStage.ABOUT_TO_SUBMIT;
 
-import java.util.Arrays;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
-import org.mockito.ArgumentCaptor;
-import org.mockito.Captor;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.mockito.junit.jupiter.MockitoSettings;
 import org.mockito.quality.Strictness;
-import uk.gov.hmcts.reform.iahomeofficeintegrationapi.domain.entities.AppealDecidedInstructMessage;
 import uk.gov.hmcts.reform.iahomeofficeintegrationapi.domain.entities.AsylumCase;
 import uk.gov.hmcts.reform.iahomeofficeintegrationapi.domain.entities.AsylumCaseDefinition;
-import uk.gov.hmcts.reform.iahomeofficeintegrationapi.domain.entities.CourtType;
-import uk.gov.hmcts.reform.iahomeofficeintegrationapi.domain.entities.MessageType;
-import uk.gov.hmcts.reform.iahomeofficeintegrationapi.domain.entities.Outcome;
 import uk.gov.hmcts.reform.iahomeofficeintegrationapi.domain.entities.ccd.Event;
 import uk.gov.hmcts.reform.iahomeofficeintegrationapi.domain.entities.ccd.callback.PreSubmitCallbackResponse;
 import uk.gov.hmcts.reform.iahomeofficeintegrationapi.domain.entities.ccd.callback.PreSubmitCallbackStage;
-import uk.gov.hmcts.reform.iahomeofficeintegrationapi.domain.service.FtpaAppealDecidedNote;
+import uk.gov.hmcts.reform.iahomeofficeintegrationapi.domain.service.FtpaDecidedNotificationsHelper;
 import uk.gov.hmcts.reform.iahomeofficeintegrationapi.domain.service.HomeOfficeInstructService;
-
 
 
 @ExtendWith(MockitoExtension.class)
@@ -48,8 +35,8 @@ class LeadershipJudgeFtpaDecidedNotificationHandlerTest extends AbstractNotifica
     @Mock
     private HomeOfficeInstructService homeOfficeInstructService;
 
-    @Captor
-    private ArgumentCaptor<AppealDecidedInstructMessage> appealDecidedInstructMessageCaptor;
+    @Mock
+    private FtpaDecidedNotificationsHelper ftpaDecidedNotificationsHelper;
 
     private LeadershipJudgeFtpaDecidedNotificationHandler leadershipJudgeFtpaDecidedNotificationHandler;
 
@@ -57,28 +44,23 @@ class LeadershipJudgeFtpaDecidedNotificationHandlerTest extends AbstractNotifica
     void setUp() {
         leadershipJudgeFtpaDecidedNotificationHandler =
             new LeadershipJudgeFtpaDecidedNotificationHandler(
-                homeOfficeInstructService, notificationsHelper
+                homeOfficeInstructService, notificationsHelper, ftpaDecidedNotificationsHelper
             );
     }
 
     @ParameterizedTest
-    @CsvSource(
-        value = {
-        "appellant:granted", "appellant:partiallyGranted", "appellant:refused", "appellant:notAdmitted",
-        "respondent:granted", "respondent:partiallyGranted", "respondent:refused", "respondent:notAdmitted"
-        },
-        delimiter = ':'
-    )
+    @ValueSource(strings = {"appellant", "respondent"})
     @MockitoSettings(strictness = Strictness.WARN)
-    void check_handler_returns_case_data_for_valid_input(String applicantType, String ftpaDecisionOutcome) {
+    void check_handler_returns_case_data_for_valid_input(String applicantType) {
 
         setupCase(Event.LEADERSHIP_JUDGE_FTPA_DECISION);
-        setupCaseData();
-        setupHelperResponses();
-        setupFtpaDecisionData(applicantType, ftpaDecisionOutcome);
 
-        when(homeOfficeInstructService.sendNotification(any(AppealDecidedInstructMessage.class)))
-            .thenReturn("OK");
+        when(asylumCase.read(AsylumCaseDefinition.FTPA_APPLICANT_TYPE, String.class))
+            .thenReturn(Optional.of(applicantType));
+
+        when(ftpaDecidedNotificationsHelper.handleFtpaDecidedNotification(
+            asylumCase, notificationsHelper, homeOfficeInstructService, Event.LEADERSHIP_JUDGE_FTPA_DECISION, "")
+        ).thenReturn("OK");
 
         PreSubmitCallbackResponse<AsylumCase> response =
             leadershipJudgeFtpaDecidedNotificationHandler.handle(ABOUT_TO_SUBMIT, callback);
@@ -87,34 +69,21 @@ class LeadershipJudgeFtpaDecidedNotificationHandlerTest extends AbstractNotifica
         assertThat(response.getData()).isNotEmpty();
         assertThat(response.getData()).isEqualTo(asylumCase);
         assertTrue(response.getErrors().isEmpty());
-
-        verify(asylumCase, times(1)).write(
-            valueOf(format("HOME_OFFICE_FTPA_%s_DECIDED_INSTRUCT_STATUS", applicantType.toUpperCase())), "OK");
-
-        verify(homeOfficeInstructService).sendNotification(appealDecidedInstructMessageCaptor.capture());
-
-        final AppealDecidedInstructMessage instructMessage = appealDecidedInstructMessageCaptor.getValue();
-
-        assertNotificationInstructMessage(instructMessage, applicantType, ftpaDecisionOutcome);
     }
 
     @ParameterizedTest
-    @CsvSource(
-        value = {
-            "appellant:granted", "appellant:partiallyGranted", "appellant:refused", "appellant:notAdmitted",
-            "respondent:granted", "respondent:partiallyGranted", "respondent:refused", "respondent:notAdmitted"
-        },
-        delimiter = ':'
-    )
+    @ValueSource(strings = {"appellant", "respondent"})
     @MockitoSettings(strictness = Strictness.WARN)
-    void check_handler_returns_error_status(String applicantType, String ftpaDecisionOutcome) {
+    void check_handler_returns_error_status(String applicantType) {
 
         setupCase(Event.LEADERSHIP_JUDGE_FTPA_DECISION);
-        setupCaseData();
-        setupHelperResponses();
-        setupFtpaDecisionData(applicantType, ftpaDecisionOutcome);
-        when(homeOfficeInstructService.sendNotification(any(AppealDecidedInstructMessage.class)))
-            .thenReturn("FAIL");
+
+        when(asylumCase.read(AsylumCaseDefinition.FTPA_APPLICANT_TYPE, String.class))
+            .thenReturn(Optional.of(applicantType));
+
+        when(ftpaDecidedNotificationsHelper.handleFtpaDecidedNotification(
+            asylumCase, notificationsHelper, homeOfficeInstructService, Event.LEADERSHIP_JUDGE_FTPA_DECISION, "")
+        ).thenReturn("FAIL");
 
         PreSubmitCallbackResponse<AsylumCase> response =
             leadershipJudgeFtpaDecidedNotificationHandler.handle(ABOUT_TO_SUBMIT, callback);
@@ -123,15 +92,6 @@ class LeadershipJudgeFtpaDecidedNotificationHandlerTest extends AbstractNotifica
         assertThat(response.getData()).isNotEmpty();
         assertThat(response.getData()).isEqualTo(asylumCase);
         assertThat(response.getErrors()).isEmpty();
-
-        verify(asylumCase, times(1)).write(
-            valueOf(format("HOME_OFFICE_FTPA_%s_DECIDED_INSTRUCT_STATUS", applicantType.toUpperCase())), "FAIL");
-
-        verify(homeOfficeInstructService).sendNotification(appealDecidedInstructMessageCaptor.capture());
-
-        final AppealDecidedInstructMessage instructMessage = appealDecidedInstructMessageCaptor.getValue();
-        assertNotificationInstructMessage(instructMessage, applicantType, ftpaDecisionOutcome);
-
     }
 
     @Test
@@ -197,43 +157,4 @@ class LeadershipJudgeFtpaDecidedNotificationHandlerTest extends AbstractNotifica
             .isExactlyInstanceOf(NullPointerException.class);
     }
 
-    @Test
-    void should_respond_fail_for_case_reference_null_value() {
-
-        setupCase(Event.LEADERSHIP_JUDGE_FTPA_DECISION);
-        String applicantType = "appellant";
-
-        when(asylumCase.read(AsylumCaseDefinition.FTPA_APPLICANT_TYPE, String.class))
-            .thenReturn(Optional.of(applicantType));
-
-        PreSubmitCallbackResponse<AsylumCase> response =
-            leadershipJudgeFtpaDecidedNotificationHandler.handle(ABOUT_TO_SUBMIT, callback);
-
-        verify(asylumCase, times(1)).write(
-            valueOf(format("HOME_OFFICE_FTPA_%s_DECIDED_INSTRUCT_STATUS", applicantType.toUpperCase())), "FAIL");
-    }
-
-    private void setupFtpaDecisionData(String applicantType, String ftpaDecisionOutcome) {
-
-        when(asylumCase.read(AsylumCaseDefinition.FTPA_APPLICANT_TYPE, String.class))
-            .thenReturn(Optional.of(applicantType));
-
-        when(asylumCase.read(
-            valueOf(format("FTPA_%s_DECISION_OUTCOME_TYPE", applicantType.toUpperCase())), String.class)
-        ).thenReturn(Optional.of(ftpaDecisionOutcome));
-    }
-
-    private void assertNotificationInstructMessage(
-        AppealDecidedInstructMessage instructMessage, String applicantType, String ftpaDecisionOutcome) {
-
-        assertThat(instructMessage.getConsumerReference()).isEqualTo(consumerReference);
-        assertThat(instructMessage.getMessageHeader()).isEqualTo(messageHeader);
-        assertThat(instructMessage.getHoReference()).isEqualTo(someDocumentReference);
-        assertThat(instructMessage.getNote())
-            .contains(FtpaAppealDecidedNote.fromId(ftpaDecisionOutcome + "_" + applicantType).getValue());
-
-        assertThat(instructMessage.getMessageType()).isEqualTo(MessageType.COURT_OUTCOME.name());
-        assertThat(instructMessage.getCourtOutcome().getCourtType()).isEqualTo(CourtType.FIRST_TIER);
-        assertThat(instructMessage.getCourtOutcome().getOutcome()).isIn(Arrays.asList(Outcome.values()));
-    }
 }
